+++ eris/src/fields/fr.rs
@@ -1,0 +1,1 @@
+pub use ark_pluto::{Fq as Fr, FqParameters as FrParameters};